data type conversions, removing unreferenced variables

+++ liboptical/optical_structures.h
@@ -72,7 +72,7 @@ struct opt_results
 
     int n_schedule;
 
-    float n_replacements;
+    int n_replacements;
 
 	float heliostat_refurbish_cost;
 	float heliostat_refurbish_cost_y1;
